Put detail view model factory in better place
