+++ app/src/main/java/test/mug/espresso/detailView/DetailViewModel.kt
@@ -3,17 +3,25 @@ package test.mug.espresso.detailView
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import test.mug.espresso.domain.PowerMug
 
 class DetailViewModel(powerMug: PowerMug) : ViewModel() {
 	private val _selectedPlace = MutableLiveData<PowerMug>()
-
-	// The external LiveData for the SelectedProperty
 	val selectedPlace: LiveData<PowerMug>
 		get() = _selectedPlace
 
-	// Initialize the _selectedProperty MutableLiveData
 	init {
 		_selectedPlace.value = powerMug
+	}
+
+	class Factory(val powerMug: PowerMug) : ViewModelProvider.Factory {
+		@Suppress("unchecked_cast")
+		override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+			if (modelClass.isAssignableFrom(DetailViewModel::class.java)) {
+				return DetailViewModel(powerMug) as T
+			}
+			throw IllegalArgumentException("Unknown ViewModel class")
+		}
 	}
 }
